requirements.txt
python-dotenv>=1.0.1
neo4j>=5.22.0
langchain>=0.2.14
langgraph>=0.2.38
langgraph-checkpoint-postgres>=2.0.0
psycopg[binary,pool]>=3.1.0

# ============================================================================
# [NEW] 2025-01-16: 完全异步Checkpointer支持（彻底修复阻塞调用问题）
# ============================================================================
langgraph-checkpoint-sqlite>=0.1.0  # 开发环境推荐，完全异步，无阻塞调用
# ============================================================================

langchain-tavily>=0.1.0
tavily-python>=0.4.0
requests>=2.32.3
openai>=1.43.0
langchain-openai>=0.1.22

# MongoDB
pymongo>=4.6.0
langchain-mongodb>=0.1.0

# Graph Database
neo4j>=5.22.0

# Vector Database & RAG
chromadb>=0.4.22
pypdf>=4.0.0
langchain-chroma>=0.1.2
langchain-community>=0.2.11
sentence-transformers>=2.3.0
langchain-milvus>=0.1.0
pymilvus>=2.4.4

# UI & Visualization
rich>=13.0.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

